requirements.txt
Markdown==2.4
pyapi-gitlab==6.1.6
requests==2.2.1
